backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
python-multipart
tensorflow
numpy
pillow
